Improve code organization of organizeDeclarations rule and related helpers
