+++ Sources/ParsingHelpers.swift
@@ -1881,116 +1881,6 @@ extension Formatter {
         return (argumentNames: argumentNames, inKeywordIndex: inKeywordIndex)
     }
 
-    enum Declaration: Equatable {
-        /// A type-like declaration with body of additional declarations (`class`, `struct`, etc)
-        indirect case type(
-            kind: String,
-            open: [Token],
-            body: [Declaration],
-            close: [Token],
-            originalRange: ClosedRange<Int>
-        )
-
-        /// A simple declaration (like a property or function)
-        case declaration(
-            kind: String,
-            tokens: [Token],
-            originalRange: ClosedRange<Int>
-        )
-
-        /// A #if ... #endif conditional compilation block with a body of additional declarations
-        indirect case conditionalCompilation(
-            open: [Token],
-            body: [Declaration],
-            close: [Token],
-            originalRange: ClosedRange<Int>
-        )
-
-        /// The tokens in this declaration
-        var tokens: [Token] {
-            switch self {
-            case let .declaration(_, tokens, _):
-                return tokens
-            case let .type(_, openTokens, bodyDeclarations, closeTokens, _),
-                 let .conditionalCompilation(openTokens, bodyDeclarations, closeTokens, _):
-                return openTokens + bodyDeclarations.flatMap { $0.tokens } + closeTokens
-            }
-        }
-
-        /// The opening tokens of the declaration (before the body)
-        var openTokens: [Token] {
-            switch self {
-            case .declaration:
-                return tokens
-            case let .type(_, open, _, _, _),
-                 let .conditionalCompilation(open, _, _, _):
-                return open
-            }
-        }
-
-        /// The body of this declaration, if applicable
-        var body: [Declaration]? {
-            switch self {
-            case .declaration:
-                return nil
-            case let .type(_, _, body, _, _),
-                 let .conditionalCompilation(_, body, _, _):
-                return body
-            }
-        }
-
-        /// The closing tokens of the declaration (after the body)
-        var closeTokens: [Token] {
-            switch self {
-            case .declaration:
-                return []
-            case let .type(_, _, _, close, _),
-                 let .conditionalCompilation(_, _, close, _):
-                return close
-            }
-        }
-
-        /// The keyword that determines the specific type of declaration that this is
-        /// (`class`, `func`, `let`, `var`, etc.)
-        var keyword: String {
-            switch self {
-            case let .declaration(kind, _, _),
-                 let .type(kind, _, _, _, _):
-                return kind
-            case .conditionalCompilation:
-                return "#if"
-            }
-        }
-
-        /// Whether or not this declaration defines a type (a class, enum, etc, but not an extension)
-        var definesType: Bool {
-            ["class", "actor", "struct", "enum", "protocol", "typealias"].contains(keyword)
-        }
-
-        /// The name of this type or variable
-        var name: String? {
-            let parser = Formatter(openTokens)
-            guard let keywordIndex = openTokens.firstIndex(of: .keyword(keyword)),
-                  let nameIndex = parser.index(of: .nonSpaceOrCommentOrLinebreak, after: keywordIndex),
-                  parser.tokens[nameIndex].isIdentifierOrKeyword
-            else {
-                return nil
-            }
-
-            return parser.fullyQualifiedName(startingAt: nameIndex).name
-        }
-
-        /// The original range of the tokens of this declaration in the original source file
-        var originalRange: ClosedRange<Int> {
-            switch self {
-            case let .type(_, _, _, _, originalRange),
-                 let .declaration(_, _, originalRange),
-                 let .conditionalCompilation(_, _, _, originalRange):
-                return originalRange
-            }
-        }
-    }
-
     /// The fully qualified name starting at the given index
     func fullyQualifiedName(startingAt index: Int) -> (name: String, endIndex: Int) {
         // If the identifier is followed by a dot, it's actually the first
@@ -2052,222 +1942,6 @@ extension Formatter {
             return [name.string]
         default:
             return nil
-        }
-    }
-
-    /// Returns the end index of the `Declaration` containing `declarationKeywordIndex`.
-    ///  - `declarationKeywordIndex.isDeclarationTypeKeyword` must be `true`
-    ///    (e.g. it must be a keyword like `let`, `var`, `func`, `class`, etc.
-    ///  - Parameter `fallBackToEndOfScope`: whether or not to return the end of the current
-    ///    scope if this is the last declaration in the current scope. If `false`,
-    ///    returns `nil` if this declaration is not followed by some other declaration.
-    func endOfDeclaration(
-        atDeclarationKeyword declarationKeywordIndex: Int,
-        fallBackToEndOfScope: Bool = true
-    ) -> Int? {
-        assert(tokens[declarationKeywordIndex].isDeclarationTypeKeyword
-            || tokens[declarationKeywordIndex] == .startOfScope("#if"))
-
-        // Get declaration keyword
-        var searchIndex = declarationKeywordIndex
-        let declarationKeyword = declarationType(at: declarationKeywordIndex) ?? "#if"
-        switch tokens[declarationKeywordIndex] {
-        case .startOfScope("#if"):
-            // For conditional compilation blocks, the `declarationKeyword` _is_ the `startOfScope`
-            // so we can immediately skip to the corresponding #endif
-            if let endOfConditionalCompilationScope = endOfScope(at: declarationKeywordIndex) {
-                searchIndex = endOfConditionalCompilationScope
-            }
-        case .keyword("class") where declarationKeyword != "class":
-            // Most declarations will include exactly one token that `isDeclarationTypeKeyword` in
-            //  - `class func` methods will have two (and the first one will be incorrect!)
-            searchIndex = index(of: .keyword(declarationKeyword), after: declarationKeywordIndex) ?? searchIndex
-        case .keyword("import"):
-            // Symbol imports (like `import class Module.Type`) will have an extra `isDeclarationTypeKeyword`
-            // immediately following their `declarationKeyword`, so we need to skip them.
-            if let symbolTypeKeywordIndex = index(of: .nonSpaceOrComment, after: declarationKeywordIndex),
-               tokens[symbolTypeKeywordIndex].isDeclarationTypeKeyword
-            {
-                searchIndex = symbolTypeKeywordIndex
-            }
-        case .keyword("protocol"), .keyword("struct"), .keyword("actor"),
-             .keyword("enum"), .keyword("extension"):
-            if let scopeStart = index(of: .startOfScope("{"), after: declarationKeywordIndex) {
-                searchIndex = endOfScope(at: scopeStart) ?? searchIndex
-            }
-        default:
-            break
-        }
-
-        // Search for the next declaration so we know where this declaration ends.
-        let nextDeclarationKeywordIndex = index(after: searchIndex, where: {
-            $0.isDeclarationTypeKeyword || $0 == .startOfScope("#if")
-        })
-
-        // Search backward from the next declaration keyword to find where declaration begins.
-        var endOfDeclaration = nextDeclarationKeywordIndex.flatMap {
-            index(before: startOfModifiers(at: $0, includingAttributes: true), where: {
-                !$0.isSpaceOrCommentOrLinebreak
-            }).map { endOfLine(at: $0) }
-        }
-
-        // Prefer keeping linebreaks at the end of a declaration's tokens,
-        // instead of the start of the next delaration's tokens
-        while let linebreakSearchIndex = endOfDeclaration,
-              token(at: linebreakSearchIndex + 1)?.isLinebreak == true
-        {
-            endOfDeclaration = linebreakSearchIndex + 1
-        }
-
-        // If there was another declaration after this one in the same scope,
-        // then we know this declaration ends before that one starts
-        if let endOfDeclaration = endOfDeclaration {
-            return endOfDeclaration
-        }
-
-        // Otherwise this is the last declaration in the scope.
-        // To know where this declaration ends we just have to know where
-        // the parent scope ends.
-        //  - We don't do this inside `parseDeclarations` itself since it handles this cases
-        if fallBackToEndOfScope,
-           declarationKeywordIndex != 0,
-           let endOfParentScope = endOfScope(at: declarationKeywordIndex - 1),
-           let endOfDeclaration = index(of: .nonSpaceOrLinebreak, before: endOfParentScope)
-        {
-            return endOfDeclaration
-        }
-
-        return nil
-    }
-
-    /// Parses all of the declarations in the file
-    func parseDeclarations() -> [Declaration] {
-        guard !tokens.isEmpty else { return [] }
-        return parseDeclarations(in: ClosedRange(0 ..< tokens.count))
-    }
-
-    /// Parses the declarations in the given range.
-    func parseDeclarations(in range: ClosedRange<Int>) -> [Declaration] {
-        var declarations = [Declaration]()
-        var startOfDeclaration = range.lowerBound
-        forEachToken(onlyWhereEnabled: false) { i, token in
-            guard range.contains(i),
-                  i >= startOfDeclaration,
-                  token.isDeclarationTypeKeyword || token == .startOfScope("#if")
-            else {
-                return
-            }
-
-            let declarationKeyword = declarationType(at: i) ?? "#if"
-            let endOfDeclaration = self.endOfDeclaration(atDeclarationKeyword: i, fallBackToEndOfScope: false)
-
-            let declarationRange = startOfDeclaration ... min(endOfDeclaration ?? .max, range.upperBound)
-            startOfDeclaration = declarationRange.upperBound + 1
-
-            declarations.append(.declaration(
-                kind: isEnabled ? declarationKeyword : "",
-                tokens: Array(tokens[declarationRange]),
-                originalRange: declarationRange
-            ))
-        }
-        if startOfDeclaration < range.upperBound {
-            let declarationRange = startOfDeclaration ..< tokens.count
-            declarations.append(.declaration(
-                kind: "",
-                tokens: Array(tokens[declarationRange]),
-                originalRange: ClosedRange(declarationRange)
-            ))
-        }
-
-        return declarations.map { declaration in
-            // Parses this declaration into a body of declarations separate from the start and end tokens
-            func parseBody(in bodyRange: Range<Int>) -> (start: [Token], body: [Declaration], end: [Token]) {
-                var startTokens = Array(tokens[declaration.originalRange.lowerBound ..< bodyRange.lowerBound])
-                var endTokens = Array(tokens[bodyRange.upperBound ... declaration.originalRange.upperBound])
-
-                guard !bodyRange.isEmpty else {
-                    return (start: startTokens, body: [], end: endTokens)
-                }
-
-                var bodyRange = ClosedRange(bodyRange)
-
-                // Move the leading newlines from the `body` into the `start` tokens
-                // so the first body token is the start of the first declaration
-                while tokens[bodyRange].first?.isLinebreak == true {
-                    startTokens.append(tokens[bodyRange.lowerBound])
-
-                    if bodyRange.count > 1 {
-                        bodyRange = (bodyRange.lowerBound + 1) ... bodyRange.upperBound
-                    } else {
-                        // If this was the last remaining token in the body, just return now.
-                        // We can't have an empty `bodyRange`.
-                        return (start: startTokens, body: [], end: endTokens)
-                    }
-                }
-
-                // Move the closing brace's indentation token from the `body` into the `end` tokens
-                if tokens[bodyRange].last?.isSpace == true {
-                    endTokens.insert(tokens[bodyRange.upperBound], at: endTokens.startIndex)
-
-                    if bodyRange.count > 1 {
-                        bodyRange = bodyRange.lowerBound ... (bodyRange.upperBound - 1)
-                    } else {
-                        // If this was the last remaining token in the body, just return now.
-                        // We can't have an empty `bodyRange`.
-                        return (start: startTokens, body: [], end: endTokens)
-                    }
-                }
-
-                // Parse the inner body declarations of the type
-                let bodyDeclarations = parseDeclarations(in: bodyRange)
-
-                return (startTokens, bodyDeclarations, endTokens)
-            }
-
-            // If this declaration represents a type, we need to parse its inner declarations as well.
-            let typelikeKeywords = ["class", "actor", "struct", "enum", "protocol", "extension"]
-
-            if typelikeKeywords.contains(declaration.keyword),
-               let declarationTypeKeywordIndex = index(
-                   in: Range(declaration.originalRange),
-                   where: { $0.string == declaration.keyword }
-               ),
-               let bodyOpenBrace = index(of: .startOfScope("{"), after: declarationTypeKeywordIndex),
-               let bodyClosingBrace = endOfScope(at: bodyOpenBrace)
-            {
-                let bodyRange = (bodyOpenBrace + 1) ..< bodyClosingBrace
-                let (startTokens, bodyDeclarations, endTokens) = parseBody(in: bodyRange)
-
-                return .type(
-                    kind: declaration.keyword,
-                    open: startTokens,
-                    body: bodyDeclarations,
-                    close: endTokens,
-                    originalRange: declaration.originalRange
-                )
-            }
-
-            // If this declaration represents a conditional compilation block,
-            // we also have to parse its inner declarations.
-            else if declaration.keyword == "#if",
-                    let declarationTypeKeywordIndex = index(
-                        in: Range(declaration.originalRange),
-                        where: { $0.string == "#if" }
-                    ),
-                    let endOfBody = endOfScope(at: declarationTypeKeywordIndex)
-            {
-                let startOfBody = endOfLine(at: declarationTypeKeywordIndex)
-                let (startTokens, bodyDeclarations, endTokens) = parseBody(in: startOfBody ..< endOfBody)
-
-                return .conditionalCompilation(
-                    open: startTokens,
-                    body: bodyDeclarations,
-                    close: endTokens,
-                    originalRange: declaration.originalRange
-                )
-            } else {
-                return declaration
-            }
         }
     }
